+++ Database/graphql.go
@@ -1,0 +1,137 @@
+package Database
+
+import "github.com/graphql-go/graphql"
+
+func defineCustomerObject() {
+	customerType = graphql.NewObject(graphql.ObjectConfig{
+		Name:        "Customer",
+		Description: "A customer of the company.",
+		Fields: graphql.Fields{
+			"id": &graphql.Field{
+				Type:        graphql.NewNonNull(graphql.Int),
+				Description: "The id of the customer.",
+				Resolve: func(p graphql.ResolveParams) (interface{}, error) {
+					// ToDo: How does this Source work exactly? O_o
+					customer, ok := p.Source.(Customer);
+					if ok {
+						return customer.id, nil
+					}
+					return nil, nil
+				},
+			},
+			"name": &graphql.Field{
+				Type:        graphql.String,
+				Description: "The name of the customer.",
+				Resolve: func(p graphql.ResolveParams) (interface{}, error) {
+					customer, ok := p.Source.(Customer);
+					if ok {
+						return customer.name, nil
+					}
+					return nil, nil
+				},
+			},
+			"surname": &graphql.Field{
+				Type:        graphql.String,
+				Description: "The name of the customer.",
+				Resolve: func(p graphql.ResolveParams) (interface{}, error) {
+					customer, ok := p.Source.(Customer);
+					if ok {
+						return customer.name, nil
+					}
+					return nil, nil
+				},
+			},"address": &graphql.Field{
+				Type:        graphql.String,
+				Description: "The name of the customer.",
+				Resolve: func(p graphql.ResolveParams) (interface{}, error) {
+					customer, ok := p.Source.(Customer);
+					if ok {
+						return customer.name, nil
+					}
+					return nil, nil
+				},
+			},"skill": &graphql.Field{
+				Type:        graphql.String,
+				Description: "The name of the customer.",
+				Resolve: func(p graphql.ResolveParams) (interface{}, error) {
+					customer, ok := p.Source.(Customer);
+					if ok {
+						return customer.name, nil
+					}
+					return nil, nil
+				},
+			},"email": &graphql.Field{
+				Type:        graphql.String,
+				Description: "The name of the customer.",
+				Resolve: func(p graphql.ResolveParams) (interface{}, error) {
+					customer, ok := p.Source.(Customer);
+					if ok {
+						return customer.name, nil
+					}
+					return nil, nil
+				},
+			},"telephone": &graphql.Field{
+				Type:        graphql.String,
+				Description: "The name of the customer.",
+				Resolve: func(p graphql.ResolveParams) (interface{}, error) {
+					customer, ok := p.Source.(Customer);
+					if ok {
+						return customer.name, nil
+					}
+					return nil, nil
+				},
+			},"birthday": &graphql.Field{
+				Type:        graphql.String,
+				Description: "The name of the customer.",
+				Resolve: func(p graphql.ResolveParams) (interface{}, error) {
+					customer, ok := p.Source.(Customer);
+					if ok {
+						return customer.name, nil
+					}
+					return nil, nil
+				},
+			},
+		},
+	})
+}
+
+func defineCustomerSchema() {
+	queryType := graphql.NewObject(graphql.ObjectConfig{
+		Name: "Query",
+		Fields: graphql.Fields{
+			"customer": &graphql.Field{
+				Type: customerType,
+				Args: graphql.FieldConfigArgument{
+					//Do we need only id?? its to get the whole user, but maybe some selections would be good at customers
+					"id": &graphql.ArgumentConfig{
+						Description: "id of the user",
+						Type:        graphql.NewNonNull(graphql.Int),
+					},
+
+				},
+				Resolve: func(p graphql.ResolveParams) (interface{}, error) {
+					id := p.Args["id"].(int)
+					//return GetUser(id), nil
+					return id, nil
+				},
+			},
+			"customers": &graphql.Field{
+				Type: graphql.NewList(customerType),
+				Resolve: func(p graphql.ResolveParams) (interface{}, error) {
+					// Users are a map, but we need a list!
+					userSlice := make([]User, len(Users))
+					idx := 0
+					for  _, user := range Users {
+						userSlice[idx] = user
+						idx++
+					}
+					return userSlice, nil
+				},
+			},
+		},
+	})
+	customerSchema, _ = graphql.NewSchema(graphql.SchemaConfig{
+		Query: queryType,
+	})
+}
+
